refactor: split organismview into shared sections with per-site main and pangenome (#1612)
