+++ app/viewModelBuilders/catalog/ga2/viewModelBuilders.tsx
@@ -3,16 +3,12 @@ import {
   type GA2OrganismEntity,
 } from "@/apis/catalog/ga2/entities";
 import { type OrganismAvatar } from "@/components/Entity/components/OrganismAvatar/organismAvatar";
-import {
-  COLUMN_PRESET_KEY,
-  COLUMN_PRESET_LABEL,
-} from "@/views/OrganismView/components/Main/constants";
-import { type Main as OrganismViewMain } from "@/views/OrganismView/components/Main/main";
 import { type BackPageHero } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/components/BackPageHero/backPageHero";
 import { type Link } from "@databiosphere/findable-ui/lib/components/Links/components/Link/link";
 import { COLUMN_IDENTIFIER } from "@databiosphere/findable-ui/lib/components/Table/common/columnIdentifier";
 import { BasicCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/BasicCell/basicCell";
 import { ChipCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/ChipCell/chipCell";
+import { type Main as OrganismViewMain } from "@ga2/views/OrganismView/components/Main/main";
 import { sanitizeEntityId } from "@repo/shared/apis/utils";
 import { AnalyzeGenome } from "@repo/shared/components/Table/components/TableCell/components/AnalyzeGenome/analyzeGenome";
 import { LevelCell } from "@repo/shared/components/Table/components/TableCell/components/LevelCell/levelCell";
@@ -32,6 +28,10 @@ import {
   formatNumber,
   getGenomeStrainText,
 } from "@repo/shared/viewModelBuilders/viewModelBuilders";
+import {
+  COLUMN_PRESET_KEY,
+  COLUMN_PRESET_LABEL,
+} from "@repo/shared/views/OrganismView/components/Main/constants";
 import {
   GA2_CATEGORY_KEY,
   GA2_CATEGORY_LABEL,
