+++ packages/shared/views/OrganismView/components/Main/components/AssembliesSection/types.ts
@@ -1,4 +1,4 @@
-import { type ColumnPreset } from "@/views/OrganismView/components/Main/types";
+import { type ColumnPreset } from "@repo/shared/views/OrganismView/components/Main/types";
 import { type RowData, type TableOptions } from "@tanstack/react-table";
 
 export interface Props<T extends RowData> {
