+++ packages/shared/views/OrganismView/components/Main/components/WorkflowsSection/workflowsSection.tsx
@@ -1,12 +1,12 @@
-import { EmptyState } from "@/views/OrganismView/components/Main/components/EmptyState/emptyState";
-import { StyledSectionTitle } from "@/views/OrganismView/components/Main/main.styles";
-import { buildOrganismWorkflows } from "@/views/OrganismView/components/Main/utils";
 import { useFeatureFlag } from "@databiosphere/findable-ui/lib/hooks/useFeatureFlag/useFeatureFlag";
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import { Stack } from "@mui/material";
 import { WorkflowCategory } from "@repo/shared/components/workflow/WorkflowCategory/workflowCategory";
 import { ROUTES } from "@repo/shared/routes/constants";
 import { getWorkflows } from "@repo/shared/services/workflows/entities";
+import { EmptyState } from "@repo/shared/views/OrganismView/components/Main/components/EmptyState/emptyState";
+import { StyledSectionTitle } from "@repo/shared/views/OrganismView/components/Main/main.styles";
+import { buildOrganismWorkflows } from "@repo/shared/views/OrganismView/components/Main/utils";
 import { type JSX } from "react";
 import { type Props } from "./types";
