+++ packages/shared/views/OrganismView/components/Main/table/components/Toolbar/types.ts
@@ -1,7 +1,7 @@
-import type { ColumnPreset } from "@/views/OrganismView/components/Main/types";
+import type { ColumnPreset } from "@repo/shared/views/OrganismView/components/Main/types";
 import type { RowData, Table } from "@tanstack/react-table";
 
 export interface Props<T extends RowData> {
-  presets: ColumnPreset[];
+  columnPresets: ColumnPreset[];
   table: Table<T>;
 }
