+++ packages/shared/views/OrganismView/components/Main/table/table.styles.ts
