+++ site-config/brc-analytics/local/entity/organism/organismMainColumn.ts
@@ -1,5 +1,5 @@
 import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
-import { Main as OrganismViewMain } from "@/views/OrganismView/components/Main/main";
+import { Main as OrganismViewMain } from "@brc/views/OrganismView/components/Main/main";
 import { BackPageContentSingleColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import { type ComponentsConfig } from "@databiosphere/findable-ui/lib/config/entities";
