+++ site-config/ga2/local/entity/organism/organismMainColumn.ts
@@ -1,11 +1,11 @@
 import { type GA2OrganismEntity } from "@/apis/catalog/ga2/entities";
 import * as V from "@/viewModelBuilders/catalog/ga2/viewModelBuilders";
-import { Main as OrganismViewMain } from "@/views/OrganismView/components/Main/main";
 import { BackPageContentSingleColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import {
   type ComponentConfig,
   type ComponentsConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
+import { Main as OrganismViewMain } from "@ga2/views/OrganismView/components/Main/main";
 
 export const organismMainColumn: ComponentsConfig = [
   {
