+++ sites/brc-analytics/views/OrganismView/components/Main/components/PangenomeSection/pangenomeSection.styles.ts
