+++ sites/brc-analytics/views/OrganismView/components/Main/components/PangenomeSection/table/columns.ts
@@ -1,7 +1,7 @@
 import type { PangenomeMember } from "@/apis/catalog/brc-analytics-catalog/common/pangenome";
-import { AssemblyCell } from "@/views/OrganismView/components/Main/table/components/TableCell/components/AssemblyCell/assemblyCell";
-import { SelectionTracksCell } from "@/views/OrganismView/components/Main/table/components/TableCell/components/SelectionTracksCell/selectionTracksCell";
-import { ViewCell } from "@/views/OrganismView/components/Main/table/components/TableCell/components/ViewCell/viewCell";
+import { AssemblyCell } from "@brc/views/OrganismView/components/Main/table/components/TableCell/components/AssemblyCell/assemblyCell";
+import { SelectionTracksCell } from "@brc/views/OrganismView/components/Main/table/components/TableCell/components/SelectionTracksCell/selectionTracksCell";
+import { ViewCell } from "@brc/views/OrganismView/components/Main/table/components/TableCell/components/ViewCell/viewCell";
 import { RefCell } from "@repo/shared/components/Table/components/TableCell/components/RefCell/refCell";
 import { type ColumnDef } from "@tanstack/react-table";
 import { renderLevel } from "./viewBuilder";
