+++ sites/brc-analytics/views/OrganismView/components/Main/components/PangenomeSection/table/viewBuilder.tsx
