+++ sites/brc-analytics/views/OrganismView/components/Main/components/PangenomeSection/types.ts
