+++ sites/brc-analytics/views/OrganismView/components/Main/table/components/TableCell/components/AssemblyCell/assemblyCell.tsx
