+++ sites/brc-analytics/views/OrganismView/components/Main/table/components/TableCell/components/ViewCell/viewCell.tsx
