+++ sites/brc-analytics/views/OrganismView/components/Tabs/constants.ts
