+++ sites/brc-analytics/views/OrganismView/components/Tabs/types.ts
