+++ sites/brc-analytics/views/OrganismView/components/Tabs/utils.ts
